Qt: MultiCart BIOS/path persistence, mouse controller diagram, code quality cleanup
- MultiCart dialog: BIOS status row, Slot A/B path persistence, fix OK not starting the game
- Fix BIOS location resolution bug in Snes9xController (wrong enum passed to doFolder)
- SNES Mouse controller diagram swapped in for Mouse+Controller port mode
- Code quality pass: static_cast/reinterpret_cast instead of C-style casts, fixed
  signed/unsigned comparisons, removed dead code (browseSlotA/B, disableAll),
  replaced deprecated SIGNAL() disconnect, const-correct EmuCanvas getters,
  shared PanelConnectHelpers.hpp to de-duplicate GeneralPanel/EmulationPanel
  connect lambdas, split EmuMainWindow::createWidgets() into per-menu helpers

+++ qt/src/DisplayPanel.cpp
@@ -10,7 +10,7 @@ DisplayPanel::DisplayPanel(EmuApplication *app_)
     setupUi(this);
 
     connect(comboBox_driver, &QComboBox::activated, [&](int index) {
-        if (driver_list.empty() || index < 0 || index >= driver_list.size())
+        if (driver_list.empty() || index < 0 || static_cast<size_t>(index) >= driver_list.size())
             return;
 
         auto display_driver = driver_list[index].second;
